Removed ymd2rdb, rdb2ymd and fmt_rdb.
The RDB time is of the format yymmdd.hhmmss. At the time of
writing TPM, the RDB database needed this format. The code
is implemented in a way that there is no way to find the actual
year from this format, since the first two digits of yyyy format
is not kept. So this works only for years between 1900-1999. The
current versions of RDB infact use all four digits of the year.
So removing this function, since RDB is not used much and this
function is not needed.