setup_path for adding build dir to sys.path.
This module is imported and run() function executed, within all
test_*.py modules. This tries to import pytpm and if that fails,
finds the directory under ../build that contains the pytpm
directory and adds it to sys.path. If that fails, then
sys.exit() is called.

This allows tests to be run after building the module, without
needing to install it.

The script nosetests that comes with nose can now be run on
all the test function modules.