Rearranged dirs for easily running tests; switched to setuptools.
The rearrangement is based on the package structure for PyEphem.

To run tests without installing the package, I can now run

  python setup-devel.py test

and if Nose is installed then

  python setup-devel.py nosetests

The same applies to setup.py.

Distutils doesn't recognize the test or nosetests commands, mentioned
above. Setuptools does have test command and a nosetests command.
But unfortunately setuptools doesn't run Cython on the name.pyx file and
thereby converting it into name.c file. To get setuptools to recognize
Cython, I added the fake_pyrex module from the lxml project. See
comments in the setup*py files for more information.

The test files are also specifed as a sub-module of the package. They
are available as the module pytpm.tests.

Running the tests will create a directory pytpm-egg.info, which was added
to .gitignore.

1. Instead of lib, the main package directory is pytpm.
2. The test directory was copied into pytpm dir and renamed tests
3. The pytpm/tests dir was converted into a package with a dummy
   __init__.py
4. Changed distutils to setuptools.
5. Added fake_pyrex from the lxml project.
6. Changed setup.py and setup-devel.py to reflect the new directory
   structure.
7. Added the test_suite setup option.