YMD accepts j=value; fixed return type for j2ymd
YMD now accepts a scalar Julian date at initialization; if this
is present then all other keywords are ignored. So order is
j=value, if not present then year=value, if not then others.

The return type for j2ymd in _tpm_times.pxd was double; changed
it to YMD.

Test YMD initialization and YMD.to_j with YMD(j=value).