DMS str() and unicode(): fmt_dms and D instead of \u00B0.
Instead of formatting myself, just use the tpm.fmt_dms
function. The reason for having my own formatting was to get
the degree symbol instead of D, for the degrees part. But decided
to keep it simple.

It will be a trivial matter to replace D with \u00B0 in any
application code.