Initialize V3 fields.
V3 class has ctype == v3.type; type is a function.

V3 initialization sets default ctype and vtype to tpm.CARTESIAN
and tpm.POS.

Attributes are accessed using __getattr__. This checks to see
if the attribute is valid for the objects ctype and vtype. Then
returns the appropriate field of v3, which is stored as self._v3.
If attribute is not valid the AttributeError exception is raised.

Tests. A class Checker for methods that check field values depending
on ctype and vtype.