Added note on behaviour of DMS.normalize() and storage of angle.
Angle is treated as (sign*dd) + mm/60.0 + ss/3600.0 and not the
usual sign * (dd+mm/60.0+ss/3600.0). Internally DMS is consistent,
for example when angle is printed out it gives the proper value.
Normalize method will set the components in the first manner. And
when components are set manually, this has to kept in mind. This
WILL lead to confusion.