Simplified setup.py.
Cython does not get run on .pyx if other C files, from TPM, are
included. But it does get run if instead of C files, the library
itself is specified as in setup-devel.py. So a lot of the code
in setup.py was redundant. Removed them. Now setup.py does not
deal with Cython at all. It compiles the Cython generated C file
with other C files and creates the library.

According to Cython website this can be done using distutils, but
then I can run tests using setup test_suite.