Changed (ra,dec) everywhere to (alpha,delta); numpydoc docstrings.
(alpha. delta) is more appropriate, since the coordinates need not
be (ra, dec).

numpydoc style docstirngs are more readable at the terminal than
regular Sphinx docstrings. They are not as feature rich in HTML
output as Sphinx but these docstrings will most likely be read at
the terminal. This is a good compromise to make.