summary_test_slalib.py: functions to summarize test results.
In the automated tests, I only test if the difference between SLALIB
and PyTPM is within some limit. But that can be contaminated by a few
bad data points. So this module  has functions to get some details of
the comparison. In here I repeat the code used in the test_slalib.py
file. But then use Numpy and Scipy for obtaining statistics.

In this commit there is one function, hipfk54() that prints summary
of comparison between FK5 to FK4 conversion using SLALIB fk524 and
pytpm (convertv6(v, s1=6, s2=5) + PM from J2000 to B1950).